requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0
sentence-transformers==2.2.2
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
PyMuPDF==1.23.8
python-docx==1.1.0
openai==1.3.0
pydantic==2.5.0
jinja2==3.1.2
requests==2.31.0
itsdangerous==2.1.2